backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
distro==1.9.0
django==6.0.2
django-environ==0.13.0
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.64.0
googleapis-common-protos==1.72.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
kombu==5.6.2
multidict==6.7.1
oauthlib==3.3.1
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
